Revert "Add interface to Elasticsearch's `search_after` API (#157)"
This reverts commit 7ff5eac599bbbf49a6df8c658927c98d33ba9ca1.
